feat: Update CKD-EPI calculation to race-free 2021 equation
The CKD-EPI calculation in `services/utils.ts` has been updated to use the 2021 race-free equation. This simplifies the input parameters by removing the `ethnicity` field and adjusts the constants according to the latest guidelines.

The `PatientDashboard` component has been updated to reflect this change by no longer passing the ethnicity to the `calculateCKDEPI` function.